pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "swingmusic"
version = "2.0.6"
description = "Swing Music"
readme = "README.md"
requires-python = ">=3.11, <=3.12"

dependencies = [
    "pillow>=11.1.0",
    "Flask>=3.1.0",
    "Flask-Cors>=3.0.10",
    "requests>=2.27.1",
    "colorgram.py>=1.2.0",
    "tqdm>=4.65.0",
    "tinytag>=2.1.1",
    "Unidecode>=1.3.6",
    "psutil>=5.9.4",
    "show-in-file-manager>=1.1.4",
    "flask-compress>=1.13",
    "tabulate>=0.9.0",
    "setproctitle>=1.3.2",
    "locust>=2.20.1",
    "watchdog>=4.0.0",
    "flask-jwt-extended>=4.6.0",
    "sqlalchemy>=2.0.31",
    "memory-profiler>=0.61.0",
    "sortedcontainers>=2.4.0",
    "xxhash>=3.4.1",
    "ffmpeg-python>=0.2.0",
    "schedule>=1.2.2",
    "pillow>=11.1.0",
    "flask-openapi3==3.0.2",
    "rapidfuzz==3.11.0",
    "pendulum>=3.0.0",
    "pystray>=0.19.5",
    "pyinstaller==6.12.0",
    "waitress>=3.0.2; sys_platform == 'win32'",
    "bjoern >=3.2.2; sys_platform != 'win32'"
]

[tool.uv]
dependency-metadata = [
    { name = "waitress", version = "3.0.2", requires-dist = [], requires-python = ">=3.11" },
    { name = "bjoern", version="3.2.2", requires-dist = [], requires-python = ">=3.11"}
]

[project.scripts]
swingmusic = "swingmusic.__main__:main"


[project.urls]
Homepage = "https://swingmx.com/"
Repository = "https://github.com/swingmx/swingmusic"
Documentation = "https://swingmx.com/guide/introduction.html"
Issues = "https://github.com/swingmx/swingmusic/issues"


[tool.setuptools.package-data]
swingmusic = ["./client/*", "**/*.txt"]